// src/types/anime.ts
export interface AnimeCard {
  id: number;
  title: string;
  image: string;
  score: number | null;
  episodes: number | null;
  status: string;
  genres: string[];
  synopsis: string | null;
  year: number | null;
  trailer?: string;
}